+++ src/pages/dashboard/components/SearchAndInviteUser/index.ts
@@ -1,0 +1,3 @@
+import SearchAndInviteUser from './SearchAndInviteUser';
+
+export { SearchAndInviteUser };
