+++ src/pages/dashboard/components/index.ts
@@ -1,5 +1,6 @@
 import { ColumnContainer } from './ColumnContainer';
 import { MenuContainer } from './MenuContainer';
 import { Profile } from './Profile';
+import { SearchAndInviteUser } from './SearchAndInviteUser';
 
-export { ColumnContainer, MenuContainer, Profile };
+export { ColumnContainer, MenuContainer, Profile, SearchAndInviteUser };
